+++ packages/react-dom/src/Link.tsx
@@ -25,6 +25,7 @@ export interface LinkProps
   onClick?: (e: React.MouseEvent<HTMLElement>) => void;
   anchor?: React.ReactType;
   children: NavigatingChildren | React.ReactNode;
+  anchorProps?: React.AnchorHTMLAttributes<HTMLAnchorElement>;
 }
 
 interface BaseLinkProps extends LinkProps {
@@ -37,6 +38,7 @@ interface LinkState {
 }
 
 let hasWarnedTo = false;
+let hasWarnedAnchorProps = false;
 
 class BaseLink extends React.Component<BaseLinkProps, LinkState> {
   removed: boolean;
@@ -102,6 +104,7 @@ class BaseLink extends React.Component<BaseLinkProps, LinkState> {
       router,
       forwardedRef,
       children,
+      anchorProps,
       ...rest
     } = this.props;
     if (process.env.NODE_ENV !== "production") {
@@ -112,8 +115,20 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
 
 <Link name="Route Name">...</Link>`);
       }
+
+      if (!hasWarnedAnchorProps && Object.keys(rest).length > 0) {
+        hasWarnedAnchorProps = true;
+        console.warn(`Deprecation warning:
+Passing additional props to a <Link> will no longer be forwarded to the rendered component in v2.
+
+Instead, please use the "anchorProps" prop to pass an object of props to be attached to the component.
+
+<Link to="Route Name" anchorProps={{ className: "test" }}>`);
+      }
     }
+
     const routeName = name || to;
+
     const Anchor: React.ReactType = anchor ? anchor : "a";
     const href: string = router.history.toHref({
       hash,
@@ -122,9 +137,14 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
       pathname: routeName ? router.route.pathname(routeName, params) : ""
     });
 
+    const additionalProps = {
+      ...rest,
+      ...anchorProps
+    };
+
     return (
       <Anchor
-        {...rest}
+        {...additionalProps}
         onClick={this.clickHandler}
         href={href}
         ref={forwardedRef}