+++ packages/react-dom/tests/Link.spec.tsx
@@ -307,6 +307,69 @@ The "to" prop should be replaced with the "name" prop. The "to" prop will be rem
     });
   });
 
+  describe("anchorProps", () => {
+    describe("additional props (deprecated)", () => {
+      it("warns when passing additional props to the <Link>", () => {
+        const realWarn = console.warn;
+        const fakeWarn = jest.fn();
+        console.warn = fakeWarn;
+
+        ReactDOM.render(
+          <Router>
+            {() => (
+              <Link to="Test" className="hi">
+                Test
+              </Link>
+            )}
+          </Router>,
+          node
+        );
+
+        expect(fakeWarn.mock.calls.length).toBe(1);
+        expect(fakeWarn.mock.calls[0][0]).toBe(`Deprecation warning:
+Passing additional props to a <Link> will no longer be forwarded to the rendered component in v2.
+
+Instead, please use the "anchorProps" prop to pass an object of props to be attached to the component.
+
+<Link to="Route Name" anchorProps={{ className: "test" }}>`);
+
+        console.warn = realWarn;
+      });
+
+      it("passes additional props to the anchor", () => {
+        ReactDOM.render(
+          <Router>
+            {() => (
+              <Link to="Test" className="hi">
+                Test
+              </Link>
+            )}
+          </Router>,
+          node
+        );
+
+        const a = node.querySelector("a");
+        expect(a.classList.contains("hi")).toBe(true);
+      });
+    });
+
+    it("passes anchorProps to the rendered anchor", () => {
+      ReactDOM.render(
+        <Router>
+          {() => (
+            <Link to="Test" anchorProps={{ className: "hi" }}>
+              Test
+            </Link>
+          )}
+        </Router>,
+        node
+      );
+
+      const a = node.querySelector("a");
+      expect(a.classList.contains("hi")).toBe(true);
+    });
+  });
+
   describe("ref", () => {
     it("returns the anchor's ref, not the link's", () => {
       const history = InMemory();